Changed the version of sqlite to the latest. It was causing issues with the dialect.
Made changes to the application.properties to make sure that the program was running the sqlite3 dialect.
Made changes to the application.properties to make sure that the db url was correct.

Changed the ID generation to Location. sqlite is having issues with normal hibernate id generations. Now is using the id_seq.