POST endpoint testing and fixing for Accommodation
Made sure mock Location was created and saved before anything else in the MockDataGenerator
Disabled the @JsonIdentityInfo for now. Seems to still work without recursion
Changed CascadeType to .MERGE from .ALL on location field in Accommodation. This helps with POST and makes sure if there is a Location with given id then that one will be set as location